refactor: add Notifications folder
