+++ src/Components/Notifications/__tests__/NotificationItem.jest.tsx
@@ -1,12 +1,9 @@
 import { screen } from "@testing-library/react"
-import {
-  getDateLabel,
-  NotificationItemFragmentContainer,
-} from "Components/NotificationItem"
 import { setupTestWrapperTL } from "DevTools/setupTestWrapper"
 import { DateTime } from "luxon"
 import { graphql } from "react-relay"
 import { NotificationItem_test_Query } from "__generated__/NotificationItem_test_Query.graphql"
+import { NotificationItemFragmentContainer } from "../NotificationItem"
 
 jest.unmock("react-relay")
 
@@ -79,35 +76,6 @@ describe("NotificationItem", () => {
 
     const imageTwo = screen.getByAltText("Artwork image of artwork one")
     expect(imageTwo).toBeInTheDocument()
-  })
-})
-
-describe("getDateLabel", () => {
-  it("returns 'Today' label", () => {
-    const result1 = getDateLabel(DateTime.utc().minus({ hours: 1 }))
-    expect(result1).toEqual("Today")
-
-    const result2 = getDateLabel(DateTime.utc().minus({ hours: 5 }))
-    expect(result2).toEqual("Today")
-
-    const result3 = getDateLabel(
-      DateTime.utc().minus({ hours: 23, minutes: 59, seconds: 59 })
-    )
-    expect(result3).toEqual("Today")
-  })
-
-  it("returns 'x days ago' label", () => {
-    const result1 = getDateLabel(DateTime.utc().minus({ days: 1 }))
-    expect(result1).toEqual("1 days ago")
-
-    const result2 = getDateLabel(DateTime.utc().minus({ days: 2 }))
-    expect(result2).toEqual("2 days ago")
-
-    const result3 = getDateLabel(DateTime.utc().minus({ days: 31 }))
-    expect(result3).toEqual("31 days ago")
-
-    const result4 = getDateLabel(DateTime.utc().minus({ days: 365 }))
-    expect(result4).toEqual("365 days ago")
   })
 })
